Update log4j to 2.17.0 + slf4j-log4j12 to 1.7.32
This commit updates log4j to a known safe version on the 2.x branch.

This is not a safety-critical upgrade for nasus; the 1.x branch of log4j
contained no vulnerability to CVE CVE-2021-44228, and CVE-2021-4104 only
applies to applications using log4j 1.x versions with JMSAppender
enabled, so there is no evidence that nasus HTTP servers were affected by these vulnerabilities.

References:
https://nvd.nist.gov/vuln/detail/CVE-2021-44228
https://nvd.nist.gov/vuln/detail/CVE-2021-4104